Adjusted types according to new is_prime function implementation

+++ Problems/1-50/Problem37/Problem37.cpp
@@ -13,7 +13,7 @@ using primes::is_prime;
 using series::primes;
 using namespace generics;
 
-using numtype = unsigned int;
+using numtype = unsigned long;
 
 
 /**
